refactor: move relationsSchema to ModulesComponent
